Change evidence-treatment to 1-n relationship instead of n-n
