+++ core/pom.xml
@@ -47,12 +47,6 @@
 
         <!-- jackson -->
         <dependency>
-                <groupId>org.codehaus.jackson</groupId>
-                <artifactId>jackson-mapper-asl</artifactId>
-                <version>1.4.2</version>
-        </dependency>
-
-        <dependency>
                 <groupId>dom4j</groupId>
                 <artifactId>dom4j</artifactId>
                 <version>1.6.1</version>
