+++ core/src/main/java/org/mskcc/cbio/oncokb/model/Evidence.java
@@ -10,7 +10,10 @@ import org.mskcc.cbio.oncokb.util.TumorTypeUtils;
 import org.mskcc.oncotree.model.TumorType;
 
 import javax.persistence.*;
-import java.util.*;
+import java.util.Date;
+import java.util.HashSet;
+import java.util.Objects;
+import java.util.Set;
 
 
 /**
@@ -226,8 +229,8 @@ public class Evidence implements java.io.Serializable {
     @Column(name = "additional_info", length = 65535)
     private String additionalInfo;
 
-    @OneToMany(fetch = FetchType.EAGER, mappedBy = "evidenceTreatmentId.evidence", cascade = CascadeType.ALL)
-    private Set<EvidenceTreatment> evidenceTreatments = new HashSet<>(0);
+    @OneToMany(fetch = FetchType.EAGER, mappedBy = "evidence", cascade = CascadeType.ALL)
+    private Set<Treatment> treatments = new HashSet<>(0);
 
     @Column(name = "known_effect")
     private String knownEffect;
@@ -360,51 +363,18 @@ public class Evidence implements java.io.Serializable {
         this.additionalInfo = additionalInfo;
     }
 
-    @Transient
     public Set<Treatment> getTreatments() {
-        if (this.evidenceTreatments != null) {
-            Set<Treatment> treatments = new HashSet<>();
-            for (EvidenceTreatment evidenceTreatment : this.evidenceTreatments) {
-                treatments.add(evidenceTreatment.getTreatment());
-            }
-            return treatments;
-        } else {
-            return null;
-        }
+        return treatments;
     }
 
     public void setTreatments(Set<Treatment> treatments) {
-        if (treatments != null) {
-            Set<EvidenceTreatment> evidenceTreatments = new HashSet<>();
-            for (Treatment treatment : treatments) {
-                EvidenceTreatment evidenceTreatment = new EvidenceTreatment();
-                evidenceTreatment.setEvidence(this);
-                evidenceTreatment.setTreatment(treatment);
-                evidenceTreatments.add(evidenceTreatment);
-            }
-            this.evidenceTreatments = evidenceTreatments;
-        }
-    }
-
-    // In current setup, one evidence only has one treatment, then we could set priority to all
-    public void setTreatmentsPriority(int priority) {
-        for (EvidenceTreatment evidenceTreatment : this.getEvidenceTreatments()) {
-            evidenceTreatment.setPriority(priority);
-        }
+        this.treatments = treatments;
     }
 
     public void setPriority(Integer priority) {
-        for (EvidenceTreatment evidenceTreatment : this.getEvidenceTreatments()) {
-            evidenceTreatment.setPriority(priority);
+        for (Treatment treatment : this.getTreatments()) {
+            treatment.setPriority(priority);
         }
-    }
-
-    public Set<EvidenceTreatment> getEvidenceTreatments() {
-        return evidenceTreatments;
-    }
-
-    public void setEvidenceTreatments(Set<EvidenceTreatment> evidenceTreatments) {
-        this.evidenceTreatments = evidenceTreatments;
     }
 
     public String getKnownEffect() {
@@ -528,7 +498,7 @@ public class Evidence implements java.io.Serializable {
         this.propagation = e.propagation;
         // make deep copy of sets
         this.alterations = new HashSet<>(e.alterations);
-        this.evidenceTreatments = new HashSet<>(e.evidenceTreatments);
+        this.treatments = new HashSet<>(e.treatments);
         this.articles = new HashSet<>(e.articles);
         this.nccnGuidelines = new HashSet<>(e.nccnGuidelines);
         this.clinicalTrials = new HashSet<>(e.clinicalTrials);
